--- a/Gemfile
+++ b/Gemfile
@@ -4,3 +4,5 @@ source 'https://rubygems.org'
 gemspec
 
 gem "mime-types", "~> 3" if RUBY_VERSION > '2'
+gem "memory_profiler"
+gem "benchmark-ips"