--- a/test/fixtures/custom_content_type_mime.db
+++ b/test/fixtures/custom_content_type_mime.db
@@ -1,0 +1,2 @@
+liquid      application/x-liquid                                                      8bit
+mp4         video/vnd.objectvideo                                                     quoted-printable