--- a/test/fixtures/custom_ext_mime.db
+++ b/test/fixtures/custom_ext_mime.db
@@ -1,0 +1,2 @@
+lua         application/x-lua                                                         8bit
+m4v         video/vnd.objectvideo                                                     quoted-printable