@SamSaffron 2019-02-19
except for the "configure" api change looks fine to me, can you update the README with updated perf numbers and instructions on how to use the new custom db?