@gmcgibbon 2019-02-19
Added docs, updated benchmark, and removed the configure method.